BSAN 726 Group Project
In this project, my team and I built a database and ran SQL queries to uncover patterns in financial fraud. We looked at things like high-risk merchants, suspicious times of day, and customer behavior. To make the insights easier to understand, we used Power BI to build a dashboard. It was a great way to apply analytics to a real-world problem and show how data can help catch fraud before it happens.